增加mybatis-plus动态数据源,升级mybatis-plus依赖版本

+++ city-fire/src/main/java/com/hq/common/utils/PageUtils.java
@@ -1,8 +1,7 @@
 
 package com.hq.common.utils;
 
-import com.baomidou.mybatisplus.plugins.Page;
-
+import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
 import java.io.Serializable;
 import java.util.List;
 
@@ -43,8 +42,8 @@ public class PageUtils implements Serializable {
 	public PageUtils(Page<?> page) {
 		this.list = page.getRecords();
 		this.totalCount = (int)page.getTotal();
-		this.pageSize = page.getSize();
-		this.currPage = page.getCurrent();
+		this.pageSize = (int)page.getSize();
+		this.currPage = (int)page.getCurrent();
 		this.totalPage = (int)page.getPages();
 	}
